Modified by review comments.
