+++ ratis-shell/src/main/java/org/apache/ratis/shell/cli/sh/command/PeerCommand.java
@@ -17,6 +17,7 @@
  */
 package org.apache.ratis.shell.cli.sh.command;
 
+import org.apache.ratis.shell.cli.Command;
 import org.apache.ratis.shell.cli.sh.peer.AddCommand;
 import org.apache.ratis.shell.cli.sh.peer.RemoveCommand;
 import org.apache.ratis.shell.cli.sh.peer.SetPriorityCommand;
@@ -31,7 +32,7 @@ import java.util.function.Function;
  */
 public class PeerCommand extends AbstractParentCommand{
 
-  private static final List<Function<Context, AbstractRatisCommand>> SUB_COMMAND_CONSTRUCTORS
+  private static final List<Function<Context, Command>> SUB_COMMAND_CONSTRUCTORS
       = Collections.unmodifiableList(Arrays.asList(AddCommand::new, RemoveCommand::new,
       SetPriorityCommand::new));
